Updated the model for the converter option
For the converter option in the URL picker data type, the latest versions of the Umbraco 10 package save a string representing the selected CLR type, while previous versions have also saved an object with a "key" property. As of this commit, the package will now instead save an object with "type" parameter indicating the selected CLR type.

This change is implemented in order to allow adding more properties to the saved object, although this is something currently not used by the package.